refactor: make TimeTempTableBody component
