+++ src/components/Main/TimeTempTable/TimeTempTable.tsx
@@ -1,43 +1,13 @@
 import React from "react";
 import {
   Table,
-  TableBody,
   TableCell,
   TableHead,
   TableRow,
 } from "@mui/material";
-import { useSelector } from "react-redux";
-import { RootState } from "../../../redux/store";
-import { getFormattedTime } from "../../../utils";
-import TimeTempTableRow from "./TimeTempTableRow";
+import TimeTempTableBody from "./TimeTempTableBody";
 
 const TimeTempTable: React.FC = () => {
-
-  const timeTemp = useSelector((state: RootState) => state.session.timeTemp);
-
-  const TempRows =
-    timeTemp &&
-    timeTemp.map((thisTimeTemp, i) => {
-      const { time, temp, addedCoals } =
-        thisTimeTemp;
-
-      const formattedTime = getFormattedTime(time);
-
-      const tempDiff = i
-        ? temp - timeTemp[i - 1].temp
-        : 0;
-
-      const props = {
-        formattedTime,
-        temp,
-        tempDiff,
-        addedCoals,
-      };
-
-      return (
-        <TimeTempTableRow key={time} {...props}/>
-      );
-    });
 
   return (
     <Table className="temp-table" aria-label="simple table">
@@ -49,7 +19,7 @@ const TimeTempTable: React.FC = () => {
           <TableCell>Coals Added</TableCell>
         </TableRow>
       </TableHead>
-      <TableBody>{TempRows}</TableBody>
+      <TimeTempTableBody />
     </Table>
   );
 };
