+++ src/components/Main/TimeTempTable/TimeTempTableBody/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TimeTempTableBody';
